move setup llm and models to own module
