+++ src/codemodder/context.py
@@ -2,7 +2,6 @@ from __future__ import annotations
 
 import itertools
 import logging
-import os
 from pathlib import Path
 from textwrap import indent
 from typing import TYPE_CHECKING, Iterator, List
@@ -16,31 +15,17 @@ from codemodder.dependency import (
     build_failed_dependency_notification,
 )
 from codemodder.file_context import FileContext
+from codemodder.llm import setup_llm_client
 from codemodder.logging import log_list, logger
 from codemodder.project_analysis.file_parsers.package_store import PackageStore
 from codemodder.project_analysis.python_repo_manager import PythonRepoManager
 from codemodder.registry import CodemodRegistry
 from codemodder.utils.timer import Timer
 
-try:
-    from openai import AzureOpenAI, OpenAI
-except ImportError:
-    OpenAI = None
-    AzureOpenAI = None
-
-
 if TYPE_CHECKING:
     from openai import OpenAI
 
     from codemodder.codemods.base_codemod import BaseCodemod
-
-
-class MisconfiguredAIClient(ValueError):
-    pass
-
-
-MODELS = ["gpt-4-turbo-2024-04-09", "gpt-4o-2024-05-13"]
-DEFAULT_AZURE_OPENAI_API_VERSION = "2024-02-01"
 
 
 class CodemodExecutionContext:
@@ -87,41 +72,7 @@ class CodemodExecutionContext:
         self.path_exclude = path_exclude
         self.max_workers = max_workers
         self.tool_result_files_map = tool_result_files_map or {}
-        self.llm_client = self._setup_llm_client()
-
-    def _setup_llm_client(self) -> OpenAI | None:
-        if not AzureOpenAI:
-            logger.info("Azure OpenAI API client not available")
-            return None
-
-        azure_openapi_key = os.getenv("CODEMODDER_AZURE_OPENAI_API_KEY")
-        azure_openapi_endpoint = os.getenv("CODEMODDER_AZURE_OPENAI_ENDPOINT")
-        if bool(azure_openapi_key) ^ bool(azure_openapi_endpoint):
-            raise MisconfiguredAIClient(
-                "Azure OpenAI API key and endpoint must both be set or unset"
-            )
-
-        if azure_openapi_key and azure_openapi_endpoint:
-            logger.info("Using Azure OpenAI API client")
-            return AzureOpenAI(
-                api_key=azure_openapi_key,
-                api_version=os.getenv(
-                    "CODEMODDER_AZURE_OPENAI_API_VERSION",
-                    DEFAULT_AZURE_OPENAI_API_VERSION,
-                ),
-                azure_endpoint=azure_openapi_endpoint,
-            )
-
-        if not OpenAI:
-            logger.info("OpenAI API client not available")
-            return None
-
-        if not (api_key := os.getenv("CODEMODDER_OPENAI_API_KEY")):
-            logger.info("OpenAI API key not found")
-            return None
-
-        logger.info("Using OpenAI API client")
-        return OpenAI(api_key=api_key)
+        self.llm_client = setup_llm_client()
 
     def add_changesets(self, codemod_name: str, change_sets: List[ChangeSet]):
         self._changesets_by_codemod.setdefault(codemod_name, []).extend(change_sets)
@@ -244,8 +195,3 @@ class CodemodExecutionContext:
             for change in changes:
                 logger.info("  - %s", change.path)
                 logger.debug("    diff:\n%s", indent(change.diff, " " * 6))
-
-    def __getattribute__(self, attr: str):
-        if (name := attr.replace("_", "-")) in MODELS:
-            return os.getenv(f"CODEMODDER_AZURE_OPENAI_{name.upper()}_DEPLOYMENT", name)
-        return super().__getattribute__(attr)
